Custom hook made for reusing logic of fetching movies
