+++ 12-Use-popcorn/src/App.js
@@ -9,6 +9,7 @@ import NumResults from './presentational/NumResults'
 import MovieList from './presentational/MovieList'
 import WatchedSummary from './presentational/WatchedSummary'
 import WatchedMoviesList from './presentational/WatchedMoviesList'
+import { useMovies } from './useMovies'
 
 /* const tempMovieData = [
   {
@@ -61,62 +62,14 @@ const KEY = '8bde1356'
 
 export default function App() {
   const [query, setQuery] = useState('')
-  const [movies, setMovies] = useState([])
-  const [isLoading, setIsLoading] = useState(false)
-  const [error, setError] = useState('')
+
   const [selectedId, setSelectedId] = useState(null)
   const [watched, setWatched] = useState(function () {
     const storedValue = localStorage.getItem('watched')
     return JSON.parse(storedValue)
   })
 
-  useEffect(
-    function () {
-      const controller = new AbortController()
-      async function fetchMovies() {
-        try {
-          setError('')
-          setIsLoading(true)
-          const res = await fetch(
-            `http://www.omdbapi.com/?apikey=${KEY}&s=${query}`,
-            { signal: controller.signal }
-          )
-
-          if (!res.ok) {
-            throw new Error('Something went wrong with fetching movies!')
-          }
-
-          const data = await res.json()
-          if (data.Response === 'False') {
-            throw new Error('Movie not found!')
-          }
-
-          setMovies(data.Search)
-          setError('')
-        } catch (err) {
-          if (err.name !== 'AbortError') {
-            setError(err.message)
-          }
-        } finally {
-          setIsLoading(false)
-        }
-      }
-
-      if (query.length < 2) {
-        setMovies([])
-        setError('')
-        return
-      }
-
-      handleCloseMovie()
-      fetchMovies()
-
-      return function () {
-        controller.abort()
-      }
-    },
-    [query]
-  )
+  const { movies, isLoading, error } = useMovies(query)
 
   useEffect(
     function () {
