+++ 12-Use-popcorn/src/useMovies.js
@@ -1,0 +1,59 @@
+import { useEffect, useState } from 'react'
+
+const KEY = '8bde1356'
+
+export function useMovies(query) {
+  const [movies, setMovies] = useState([])
+  const [isLoading, setIsLoading] = useState(false)
+  const [error, setError] = useState('')
+
+  useEffect(
+    function () {
+      const controller = new AbortController()
+      async function fetchMovies() {
+        try {
+          setError('')
+          setIsLoading(true)
+          const res = await fetch(
+            `http://www.omdbapi.com/?apikey=${KEY}&s=${query}`,
+            { signal: controller.signal }
+          )
+
+          if (!res.ok) {
+            throw new Error('Something went wrong with fetching movies!')
+          }
+
+          const data = await res.json()
+          if (data.Response === 'False') {
+            throw new Error('Movie not found!')
+          }
+
+          setMovies(data.Search)
+          setError('')
+        } catch (err) {
+          if (err.name !== 'AbortError') {
+            setError(err.message)
+          }
+        } finally {
+          setIsLoading(false)
+        }
+      }
+
+      if (query.length < 2) {
+        setMovies([])
+        setError('')
+        return
+      }
+
+      //   handleCloseMovie()
+      fetchMovies()
+
+      return function () {
+        controller.abort()
+      }
+    },
+    [query]
+  )
+
+  return { movies, isLoading, error }
+}
